# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Riviera Building Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    patch:
      operationId: updateContract
      tags: [contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Receipt Vouchers ────────────────────────────────────────────────────────
...
  /bank-accounts:
    get:
      operationId: listBankAccounts
      tags: [bankAccounts]
      responses:
        "200":
          description: Bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bankAccounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}:
    get:
      operationId: getBankAccount
      tags: [bankAccounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bank account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
        "404":
          description: Not found
    patch:
      operationId: updateBankAccount
      tags: [bankAccounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    delete:
      operationId: deleteBankAccount
      tags: [bankAccounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Cheques ─────────────────────────────────────────────────────────────────
  /cheques:
    get:
      operationId: listCheques
      tags: [cheques]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [incoming, outgoing]
      responses:
        "200":
          description: Cheques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cheque"
    post:
      operationId: createCheque
      tags: [cheques]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChequeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cheque"
  /cheques/{id}:
    get:
      operationId: getCheque
      tags: [cheques]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cheque
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cheque"
        "404":
          description: Not found
    patch:
      operationId: updateCheque
      tags: [cheques]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChequeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cheque"
    delete:
      operationId: deleteCheque
      tags: [cheques]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Account Statements ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
        createdAt:
          type: string
...
    Contract:
      type: object
      required: [id, contractNumber, tenantId, unitId, startDate, endDate, rentAmount, currency, status]
      properties:
        id:
          type: integer
        contractNumber:
          type: string
        tenantId:
          type: integer
        tenantName:
          type: string
        unitId:
          type: integer
        unitNumber:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        rentAmountILS:
          type: number
        paymentFrequency:
          type: string
          enum: [monthly, quarterly, yearly]
        status:
          type: string
          enum: [active, expired, terminated]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ContractInput:
      type: object
      required: [tenantId, unitId, startDate, endDate, rentAmount, currency, paymentFrequency]
      properties:
        tenantId:
          type: integer
        unitId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        paymentFrequency:
          type: string
          enum: [monthly, quarterly, yearly]
        notes:
          type: ["string", "null"]
    ContractUpdate:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        paymentFrequency:
          type: string
          enum: [monthly, quarterly, yearly]
        status:
          type: string
          enum: [active, expired, terminated]
        notes:
          type: ["string", "null"]

    # ─── Receipt Vouchers ──────────────────────────────────────────────────────
...
    BankAccount:
      type: object
      required: [id, bankName, accountNumber, accountName, currency, balanceILS]
      properties:
        id:
          type: integer
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        currency:
          type: string
          enum: [ILS, USD, JOD]
        balanceILS:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BankAccountInput:
      type: object
      required: [bankName, accountNumber, accountName, currency]
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        currency:
          type: string
          enum: [ILS, USD, JOD]
        notes:
          type: ["string", "null"]
    BankAccountUpdate:
      type: object
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        currency:
          type: string
          enum: [ILS, USD, JOD]
        notes:
          type: ["string", "null"]

    # ─── Cheques ───────────────────────────────────────────────────────────────
    Cheque:
      type: object
      required: [id, chequeNumber, type, amount, currency, amountILS, bankName, chequeDate, dueDate, status, drawerName]
      properties:
        id:
          type: integer
        chequeNumber:
          type: string
        type:
          type: string
          enum: [incoming, outgoing]
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        bankName:
          type: string
        chequeDate:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, deposited, cleared, bounced, cancelled]
        drawerName:
          type: string
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ChequeInput:
      type: object
      required: [chequeNumber, type, amount, currency, exchangeRate, amountILS, bankName, chequeDate, dueDate, drawerName]
      properties:
        chequeNumber:
          type: string
        type:
          type: string
          enum: [incoming, outgoing]
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        bankName:
          type: string
        chequeDate:
          type: string
        dueDate:
          type: string
        drawerName:
          type: string
        tenantId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    ChequeUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, deposited, cleared, bounced, cancelled]
        notes:
          type: ["string", "null"]
        dueDate:
          type: string

    # ─── Account Statements ────────────────────────────────────────────────────
...